// EventFrontend2.O/src/app/core/models/models.ts
export interface pageSelection {
  skip: number;
  limit: number;
}
export interface apiResultFormat {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  data: any;
  totalData: number;
}
// ...
export interface tax {
  id: number;
  date: string | number;
  economic_zone: string;
  tax: string;
  tax_rate: number;
  net: number;
  gross: number;
  amount: number;
}
// ...
export interface editcreditnotes {
  sNo: number;
  editcreditnotes: number;
  product: number;
  quantity: number;
  unit: string;
  rate: number;
  discount: number;
  tax: number;
  amount: number;
}